Split tagstore into multiple source files
This does not actually change any of the implementations, but separates
each struct into its own file with its associated methods.
